25/03 - 21:58

+++ src/net/weg/topcar/dao/BancoVeiculos.java
@@ -8,7 +8,7 @@ import net.weg.topcar.model.veiculos.Veiculo;
 import java.util.Collections;
 import java.util.List;
 
-public class BancoVeiculos implements IBanco<Veiculo, Integer> {
+public class BancoVeiculos implements IBanco<Veiculo, Long> {
 
     private List<Veiculo> veiculos;
 
@@ -18,9 +18,9 @@ public class BancoVeiculos implements IBanco<Veiculo, Integer> {
     }
 
     @Override
-    public Veiculo buscarUm(Integer codigo) throws VeiculoNaoEncontradoException {
+    public Veiculo buscarUm(Long codigo) throws VeiculoNaoEncontradoException {
         for (Veiculo veiculo : veiculos) {
-            if (veiculo.getCODIGO()  == codigo) {
+            if (veiculo.getCODIGO() == codigo) {
                 return veiculo;
             }
         }
@@ -38,7 +38,7 @@ public class BancoVeiculos implements IBanco<Veiculo, Integer> {
     }
 
     @Override
-    public void remover(Integer codigo) throws VeiculoNaoEncontradoException {
+    public void remover(Long codigo) throws VeiculoNaoEncontradoException {
         try {
             Veiculo veiculo = buscarUm(codigo);
             veiculos.remove(veiculo);
@@ -49,7 +49,7 @@ public class BancoVeiculos implements IBanco<Veiculo, Integer> {
     }
 
     @Override
-    public Boolean existe(Integer codigo) {
+    public Boolean existe(Long codigo) {
         try {
             buscarUm(codigo);
             return true;
@@ -59,7 +59,7 @@ public class BancoVeiculos implements IBanco<Veiculo, Integer> {
     }
 
     @Override
-    public void alterar(Integer codigo, Veiculo novoVeiculo) throws VeiculoNaoEncontradoException {
+    public void alterar(Long codigo, Veiculo novoVeiculo) throws VeiculoNaoEncontradoException {
         Veiculo veiculo = buscarUm(codigo);
         veiculos.set(veiculos.indexOf(veiculo), novoVeiculo);
     }
