+++ src/net/weg/topcar/model/exceptions/VeiculoExistenteException.java
@@ -1,7 +1,7 @@
 package net.weg.topcar.model.exceptions;
 
 public class VeiculoExistenteException extends ObjetoExistenteException{
-    public VeiculoExistenteException(Integer codigo) {
+    public VeiculoExistenteException(Long codigo) {
         super("já existe um veículo com código " + codigo + "!");
     }
 }
