+++ src/net/weg/topcar/model/exceptions/VeiculoNaoEncontradoException.java
@@ -2,7 +2,7 @@ package net.weg.topcar.model.exceptions;
 
 public class VeiculoNaoEncontradoException extends ObjetoNaoEncontradoException{
 
-    public VeiculoNaoEncontradoException(Integer codigo) {
+    public VeiculoNaoEncontradoException(Long codigo) {
         super("não existe um veículo com código " + codigo + "!");
     }
 }
